combines queryrange series & value files Signed-off-by: Owen Diehl <ow.diehl@gmail.com>
